+++ examples/vanilla/stack/app/server-functions.tsx
@@ -1,0 +1,4 @@
+export async function world() {
+	"use server";
+	console.log("hello");
+}