+++ packages/vinxi-server-functions/server-handler.js
@@ -11,32 +11,28 @@ export async function handleServerAction(event) {
 		// This is the client-side case
 		const [filepath, name] = serverReference.split("#");
 		const action = (await import.meta.env.MANIFEST["server"].chunks[filepath].import())[name];
-		// Validate that this is actually a function we intended to expose and
-		// not the client trying to invoke arbitrary functions. In a real app,
-		// you'd have a manifest verifying this before even importing it.
-		// if (action.$$typeof !== Symbol.for("react.server.reference")) {
-		// 	throw new Error("Invalid action");
-		// }
-		const result = action.apply(null, await toWebRequest(event).json());
+		const text = await new Promise((resolve) => {
+			const requestBody = [];
+			event.node.req.on("data", (chunks) => {
+				console.log(chunks);
+				requestBody.push(chunks);
+			});
+			event.node.req.on("end", () => {
+				resolve(requestBody.join(""));
+			});
+		});
+		const json = JSON.parse(text);
+		const result = action.apply(null, json);
 		try {
 			// Wait for any mutations
 			const response = await result;
-			// const stream = renderToPipeableStream(result);
-			// // @ts-ignore
-			// stream._read = () => {};
-			// // @ts-ignore
-			// stream.on = (event, listener) => {
-			// 	events[event] = listener;
-			// };
 			event.node.res.setHeader("Content-Type", "application/json");
 			event.node.res.setHeader("Router", "server");
 
 			return JSON.stringify(response ?? null);
 		} catch (x) {
-			// We handle the error on the client
+			console.error(x);
 		}
-		// Refresh the client and return the value
-		// return {};
 	} else {
 		throw new Error("Invalid request");
 	}