+++ packages/vinxi/lib/load-app.js
@@ -1,14 +1,16 @@
 import { loadConfig } from "c12";
-import chokidar from "chokidar";
+import { fileURLToPath } from "url";
 
 import { createApp } from "./app.js";
+import { log } from "./logger.js";
 
 /**
  *
  * @param {string | undefined} configFile
  * @returns {Promise<import("./app.js").App>}
  */
-export async function loadApp(configFile = undefined) {
+export async function loadApp(configFile = undefined, args = {}) {
+	const stacks = typeof args.s === "string" ? [args.s] : args.s ?? [];
 	/** @type {{ config: import("./app.js").App }}*/
 	try {
 		let { config: app } = await loadConfig(
@@ -39,12 +41,19 @@ export async function loadApp(configFile = undefined) {
 			});
 
 			if (config.config) {
-				console.warn("Found vite.config.js with app config");
+				log("Found vite.config.js with app config");
 				// @ts-expect-error trying to send c12's config as app
+				//
 				return config;
 			} else {
-				console.warn("No app config found. Assuming SPA app.");
-				return createApp({
+				log("No app config found. Assuming SPA app.");
+
+				if (stacks.length) {
+					log("Applying stacks:", ...stacks);
+					return applyStacks(createApp({}), stacks);
+				}
+
+				const app = createApp({
 					routers: [
 						{
 							name: "public",
@@ -60,13 +69,38 @@ export async function loadApp(configFile = undefined) {
 						},
 					],
 				});
+
+				return applyStacks(app, stacks);
 			}
 		}
 
-		return app;
+		return applyStacks(app, stacks);
 	} catch (e) {
 		console.error(e);
 
 		return undefined;
 	}
+}
+
+async function applyStacks(app, s) {
+	const { default: resolve } = await import("resolve");
+	const stacks = await Promise.all(
+		s.map(async (stack) => {
+			if (stack.includes("/") || stack.includes("@")) {
+				var res = resolve.sync(stack, { basedir: app.config.root });
+				const mod = await import(res);
+				return mod.default;
+			}
+			const mod = await import(
+				fileURLToPath(new URL(`../stack/${stack}.js`, import.meta.url))
+			);
+			return mod.default;
+		}),
+	);
+
+	for (const stack of stacks) {
+		await app.stack(stack);
+	}
+
+	return app;
 }