+++ packages/vinxi/stack/index.js
@@ -1,0 +1,5 @@
+export function createStack(
+	/** @type {(app: import('./app.js').App) => import('./app.js').App | void}*/ stack,
+) {
+	return stack;
+}