+++ packages/vinxi/stack/server.js
@@ -1,0 +1,10 @@
+import { createStack } from "./index.js";
+
+export default createStack((app) => {
+	app.addRouter({
+		name: "server",
+		mode: "handler",
+		handler: "./index.ts",
+		target: "server",
+	});
+});