+++ examples/react/rsc/fw/CHANGELOG.md
@@ -4,6 +4,23 @@
 
 ### Patch Changes
 
+- Updated dependencies [8d1f0dd]
+  - vinxi@0.0.45
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [b0029a2]
+- Updated dependencies [cbd8d4c]
+- Updated dependencies [0cc2f4a]
+  - vinxi@0.0.44
+  - @vinxi/react-server@0.0.27
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [978b04a]
   - vinxi@0.0.43