+++ examples/react/spa/tanstack-router/CHANGELOG.md
@@ -4,6 +4,25 @@
 
 ### Patch Changes
 
+- Updated dependencies [8d1f0dd]
+  - vinxi@0.0.45
+  - @vinxi/openapi@0.0.3
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [b0029a2]
+- Updated dependencies [cbd8d4c]
+- Updated dependencies [0cc2f4a]
+  - vinxi@0.0.44
+  - @vinxi/react@0.0.12
+  - @vinxi/openapi@0.0.3
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [978b04a]
   - vinxi@0.0.43
   - @vinxi/openapi@0.0.3