+++ examples/react/spa/wouter/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [0cc2f4a]
+  - @vinxi/react@0.0.12
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [bc82d8e]
   - @vinxi/react@0.0.11